Fix warnings of strict flags

+++ sample/sample_benchmark_std_couroutine_callable_recursive.cpp
@@ -90,7 +90,7 @@ static void benchmark_round(int index) {
   long long real_switch_times = static_cast<long long>(max_task_number) * recursive_count;
 
   for (int i = 0; i < max_task_number; ++i) {
-    g_benchmark_generator_list[i].get_context()->set_value(i);
+    g_benchmark_generator_list[static_cast<size_t>(i)].get_context()->set_value(i);
   }
 
   end_time = time(nullptr);
