+++ src/libcopp/coroutine/coroutine_context.cpp
@@ -309,7 +309,8 @@ LIBCOPP_COPP_API int coroutine_context::create(coroutine_context *p, callback_ty
     return COPP_EC_ARGS_ERROR;
   }
 
-  size_t this_offset = reinterpret_cast<unsigned char *>(callee_stack.sp) - reinterpret_cast<unsigned char *>(p);
+  size_t this_offset =
+      static_cast<size_t>(reinterpret_cast<unsigned char *>(callee_stack.sp) - reinterpret_cast<unsigned char *>(p));
   if (this_offset < sizeof(coroutine_context) + private_buffer_size || this_offset > stack_offset) {
     return COPP_EC_ARGS_ERROR;
   }
